Comunicação do cliente com o serviço
+ atualização do cliente para W8.1;
+ adição de lógica para capturar os dados das páginas e submeter ao serviço;
+ adição de novo serviço com conexão com o banco atraves de LinqToSQL;

+++ GerenciadorDeEventos/GerenciadorDeEventosWCF/GerenciadorDeEventosWCF/Service1.svc.cs
@@ -83,12 +83,12 @@ namespace GerenciadorDeEventosWCF
         //
         // Eventos
         //
-        public void EventoAdicionar(string nome, string descricao, DateTime data, string numeroResponsavel, long latitude, long longitude)
+        public void EventoAdicionar(string nome, string descricao, DateTime data, string numeroResponsavel, double latitude, double longitude)
         {
             Evento evento = new Evento(nome, descricao, data, numeroResponsavel, latitude, longitude);
         }
 
-        public void EventoAtualizar(int evento_id, string nome, string descricao, DateTime data, long latitude, long longitude)
+        public void EventoAtualizar(int evento_id, string nome, string descricao, DateTime data, double latitude, double longitude)
 		{
 			Evento evento = Evento.Find(evento_id);
 			if (evento != null)
